deployer-apps/citizen-dev7/src/requirements.txt
anthropic==0.40.0
openai==1.52.0
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
typer==0.12.5
rich==13.8.1
click==8.1.7
httpx==0.27.2
pypdf==5.1.0
requests==2.31.0

